Get Notif Flow Working & Refactor Code - Pull prisma user/subscription logic into own service - Create user with phone number if it doesn't exist when code verified - Some things ESLint was yelling about
